chore(quantic) moved date and time utils to their own file (#7513)
