+++ packages/quantic/force-app/main/default/lwc/quanticUtils/__tests__/quanticUtils.test.js
@@ -2,7 +2,6 @@ import {
   I18nUtils,
   buildTemplateTextFromResult,
   copyToClipboard,
-  TimeSpan,
 } from 'c/quanticUtils';
 
 describe('c/quanticUtils', () => {
@@ -199,25 +198,6 @@ describe('c/quanticUtils', () => {
       ).toBe(
         `<a href="${testResult.clickUri}">${testResult.title}</a><br/><br/><quote></quote>`
       );
-    });
-  });
-
-  describe('TimeSpan', () => {
-    describe('getYoutubeFormatTimestamp', () => {
-      it('should return the correct YouTube timestamp format for timestamps of less than a minute', () => {
-        const timeSpan = new TimeSpan(45000); // 45 seconds
-        expect(timeSpan.getYoutubeFormatTimestamp()).toBe('0:45');
-      });
-
-      it('should return the correct YouTube timestamp format for timestamps of more than an hour', () => {
-        const timeSpan = new TimeSpan(3661000); // 1 hour, 1 minute, and 1 second
-        expect(timeSpan.getYoutubeFormatTimestamp()).toBe('1:01:01');
-      });
-
-      it('should return the correct YouTube timestamp format for timestamps of less than an hour', () => {
-        const timeSpan = new TimeSpan(300000); // 5 minutes
-        expect(timeSpan.getYoutubeFormatTimestamp()).toBe('05:00');
-      });
     });
   });
 });
